Replaced installer customization with use of environment variables
As little configuration as possible should be done in the installer, since it
makes it makes this program more difficult to package properly. It should also
be possible to use the program checked out directly from git without running the
installer. Configuration can instead be done with flags and environment
variables, and potentially a config file in the future. Also, don't hardcode eog
as the default previewer; just use xdg-open instead. The user's XDG defaults
should be respected.